Fixed mypy issues
